Fuse CompositorGlobal and CompositorHandler
